requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
python-dotenv
requests
responses
loguru
django-redis
redis
pytest==8.0.0
pytest-django==4.7.0
